remove geopandas dependency
